update pairing client public interface
