+++ Sources/WalletConnectPairing/Services/Common/DeletePairingService.swift
